src/python/tools/requirements.txt
# requirements.txt
python-telegram-bot==21.0